Fix for SerializedLocationTest.kt and SerializedLocation.kt

+++ src/main/kotlin/me/ebonjaeger/perworldinventory/serialization/LocationSerializer.kt
@@ -17,9 +17,9 @@ object LocationSerializer
     {
         val obj = JSONObject()
         obj["world"] = location.world.name
-        obj["x"] = location.x
-        obj["y"] = location.y
-        obj["z"] = location.z
+        obj["x"] = location.x.toFloat()
+        obj["y"] = location.y.toFloat()
+        obj["z"] = location.z.toFloat()
         obj["pitch"] = location.pitch
         obj["yaw"] = location.yaw
